refactor: game内オブジェクトのファクトリを別ファイルに分割した
