+++ frontend/src/features/game/Game.tsx
@@ -6,62 +6,11 @@ import { socket } from '@/socket';
 import { useSocket } from '@/hooks/useSocket';
 import { Container } from '@/components/Layout/Container';
 
-import { Ball, GameObjects, Paddle } from './Types';
-import { BallDto, GameDto, PaddleDto } from './dto/GameDto';
+import { Ball, Paddle } from './Types';
+import { GameDto } from './dto/GameDto';
 import { Keys } from './Keys';
 import { useKeyInput } from './useKeyinput';
-
-const CreateBall = (
-  ballDto: BallDto,
-  canvasWidth: number,
-  canvasHeight: number,
-): Ball => {
-  return {
-    x: ballDto.x * canvasWidth,
-    y: ballDto.y * canvasHeight,
-    radius: 10,
-  };
-};
-
-const CreatePaddle = (
-  paddleDto: PaddleDto,
-  canvasWidth: number,
-  canvasHeight: number,
-): Paddle => {
-  return {
-    x: paddleDto.x * canvasWidth,
-    y: paddleDto.y * canvasHeight,
-    width: paddleDto.width * canvasWidth,
-    height: paddleDto.height * canvasHeight,
-  };
-};
-
-const CreateGameObjects = (
-  gameDto: GameDto,
-  canvasWidth: number,
-  canvasHeight: number,
-): GameObjects => {
-  const ball = CreateBall(gameDto.ball, canvasWidth, canvasHeight);
-
-  const leftPaddle = CreatePaddle(
-    gameDto.leftPaddle,
-    canvasWidth,
-    canvasHeight,
-  );
-
-  const rightPaddle = CreatePaddle(
-    gameDto.rightPaddle,
-    canvasWidth,
-    canvasHeight,
-  );
-
-  return {
-    ball,
-    leftPaddle,
-    rightPaddle,
-    scores: gameDto.scores,
-  };
-};
+import { CreateGameObjects } from './Factory';
 
 const DrawBall = (ctx: CanvasRenderingContext2D, ball: Ball) => {
   ctx.beginPath();
